blog-backend/pyproject.toml
[tool.poetry]
name = "blog-backend"
version = "0.1.0"
description = "Backend for the blog app created with svelte and django"
authors = ["MDS <manik6418@gmail.com>"]
readme = "README.md"
packages = [{include = "blog_backend"}]

[tool.poetry.dependencies]
python = "3.10.11"
django = "^4.2"
psycopg = "^3.1.8"
django-dotenv = "^1.4.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
pillow = "^9.5.0"
gunicorn = "^20.1.0"

[tool.poetry.group.dev.dependencies]
poetry = "^1.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

